requirement.txt
Appium-Python-Client==0.28
asn1crypto==0.24.0
blinker==1.4
brotlipy==0.7.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
h11==0.7.0
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
idna==2.7
kaitaistruct==0.8
ldap3==2.5.1
mitmproxy==4.0.4
mysql-connector-python==8.0.12
passlib==1.7.1
protobuf==3.6.1
pyasn1==0.4.4
pycparser==2.18
pydivert==2.1.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyperclip==1.6.4
PySocks==1.6.8
redis==2.10.6
requests==2.19.1
ruamel.yaml==0.15.65
selenium==3.14.0
six==1.11.0
sortedcontainers==2.0.5
tornado==5.1
urllib3==1.23
urwid==2.0.1
wsproto==0.11.0
